+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "next-notion-portfolio",
-    "version": "1.0.8",
+    "version": "1.1.0",
     "private": true,
     "scripts": {
         "dev": "next dev",
@@ -19,6 +19,7 @@
         "next-compose-plugins": "^2.2.1",
         "next-seo": "^5.1.0",
         "react": "^17.0.2",
+        "react-device-detect": "^2.1.2",
         "react-dom": "17.0.2",
         "sass": "^1.32.5",
         "sharp": "^0.29.3"
