+++ public/posts.json
@@ -1,2 +1,1 @@
-[{"id":"77bb8959-895a-4481-80bc-31c0f327341b","title":"Building my portfolio blog with NextJS, Tailwind and Framer Motion using Notion as CMS","modifiedDate":"2022-03-11T19:04:00.000Z","description":"How I built my portfolio cum blog with NextJS that uses Notion as a CMS","url":"Building-my-portfolio-blog-with-NextJS-Tailwind-and-Framer-Motion-using-Notion-as-CMS","link":"","cover":{"type":"external","external":{"url":"https://images.unsplash.com/photo-1498050108023-c5249f4df085?ixlib=rb-1.2.1&q=85&fm=jpg&crop=entropy&cs=srgb"}},"tags":[{"id":"4f046e44-be62-4b79-9fee-bb784e417f98","name":"nextjs","color":"gray"},{"id":"81564360-e77c-480d-92d4-b9c07500de70","name":"tailwindcss","color":"default"},{"id":"494314da-eec2-4520-9366-fd011b32b19c","name":"typescript","color":"default"}],"publishDate":"2022-03-11T19:04:00.000Z"},{"id":"270bd300-31fc-4193-86e2-5ebd7f0de8de","title":"TestPage","modifiedDate":"2022-02-19T15:56:00.000Z","description":"A page to test all components","url":"TestPage","link":"https://developers.google.com/search/docs/advanced/structured-data/article#logo-guidelines","cover":null,"tags":[{"id":"bb082cfd-25c9-407c-99ad-52eea7b3cd32","name":"testing","color":"red"}],"publishDate":"2022-02-18"},{"id":"848205a5-0afd-4868-82e9-8934671ded82","title":"beginners guide to remote procedure calls with gRPC","modifiedDate":"2022-01-28T17:30:00.000Z","description":"A Beginners guide to gRPC.","url":"beginners-guide-to-remote-procedure-calls-with-gRPC","link":"","cover":null,"tags":[{"id":"485dae07-c0d7-4b34-a4d9-4e6c1511a064","name":"communication","color":"orange"},{"id":"ae2a0472-f4f5-49fc-9932-f5edc0c52ae8","name":"cloud-native","color":"yellow"}],"publishDate":"2022-01-28T17:30:00.000Z"}]
-
+[{"id":"10cfbb77-c9f0-4e18-afce-fbd59c990132","title":"Building a Developer Portfolio: Using Notion as a Headless CMS","modifiedDate":"2022-03-14T15:48:00.000Z","description":"Part 2 of the series breaking down my portfolio/blog development. Here I’ll break down on how I am using Notion as an headless CMS.","url":"Building-a-Developer-Portfolio-Using-Notion-as-a-Headless-CMS","link":"","cover":null,"tags":[{"id":"4f046e44-be62-4b79-9fee-bb784e417f98","name":"nextjs","color":"gray"},{"id":"54d6d94e-19ca-405d-8585-2902f3cb6d1f","name":"portfolio","color":"default"},{"id":"ab2f15df-e1fc-4d9c-807c-d18539ddd467","name":"notion","color":"default"},{"id":"494314da-eec2-4520-9366-fd011b32b19c","name":"typescript","color":"default"}],"publishDate":"2022-03-14T15:48:00.000Z"},{"id":"270bd300-31fc-4193-86e2-5ebd7f0de8de","title":"TestPage","modifiedDate":"2022-03-14T07:39:00.000Z","description":"A page to test all components","url":"TestPage","link":"https://developers.google.com/search/docs/advanced/structured-data/article#logo-guidelines","cover":{"type":"external","external":{"url":"https://dev-to-uploads.s3.amazonaws.com/uploads/articles/695aqm90a50u0g3zo0fm.png"}},"tags":[{"id":"bb082cfd-25c9-407c-99ad-52eea7b3cd32","name":"testing","color":"red"}],"publishDate":"2022-02-18"},{"id":"9823bc8e-24a9-43ec-9343-547034e6aeae","title":"A beginners guide to gRPC","modifiedDate":"2022-03-14T05:31:00.000Z","description":"This article breaks down gRPC, a RPC framework developed by Google and in use at some of the leading tech giants.","url":"A-beginners-guide-to-gRPC","link":"","cover":null,"tags":[{"id":"68d8e6ac-99c8-414b-a81c-d49c66606879","name":"grpc","color":"green"},{"id":"ae2a0472-f4f5-49fc-9932-f5edc0c52ae8","name":"cloud-native","color":"yellow"}],"publishDate":"2022-03-14T05:31:00.000Z"},{"id":"8ac874b2-46da-40da-94a7-eab069a313e9","title":"Using ansible to orchestrate your GKE cluster","modifiedDate":"2022-03-14T05:30:00.000Z","description":"Manually managing all your google cloud platform tools and settings can be really tricky. This article demonstrates how to use ansible to orchestrate and manage a GKE cluster","url":"Using-ansible-to-orchestrate-your-GKE-cluster","link":"","cover":null,"tags":[{"id":"613c971f-0dfd-4292-8d67-8f54541e7802","name":"ansible","color":"purple"},{"id":"43949a39-19d8-4355-b567-cbd79bff4323","name":"kubernetes","color":"blue"},{"id":"ffae7d6b-4e4b-42c8-bfd8-67c79d55f3fc","name":"gcp","color":"yellow"}],"publishDate":"2022-03-14T05:30:00.000Z"},{"id":"dc0a5f30-3398-434a-8e9e-c5ba0fe317aa","title":"How I prepared for my KCNA certification exam","modifiedDate":"2022-03-14T05:30:00.000Z","description":"KCNA is a beginner level certification offered by the linux foundation. This article provides my study plan for clearing KCNA","url":"How-I-prepared-for-my-KCNA-certification-exam","link":"","cover":null,"tags":[{"id":"43949a39-19d8-4355-b567-cbd79bff4323","name":"kubernetes","color":"blue"},{"id":"c7098092-4ba0-4cd5-8745-26df0805c1e4","name":"container","color":"pink"},{"id":"15b07ff6-8ea1-4f54-ad14-cf742e6b5f6f","name":"certification","color":"yellow"},{"id":"ae2a0472-f4f5-49fc-9932-f5edc0c52ae8","name":"cloud-native","color":"yellow"}],"publishDate":"2022-03-14T05:30:00.000Z"},{"id":"79e5991a-c721-4eb9-9236-5c6191677a19","title":"Building a Developer Portfolio: Improving your core web vitals with the help of Lighthouse","modifiedDate":"2022-03-14T05:24:00.000Z","description":"Part 3 of the series breaking down my portfolio/blog development. We’ll look at understanding core web vitals and improving them with the help of lighthouse","url":"Building-a-Developer-Portfolio-Improving-your-core-web-vitals-with-the-help-of-Lighthouse","link":"","cover":null,"tags":[{"id":"4f046e44-be62-4b79-9fee-bb784e417f98","name":"nextjs","color":"gray"},{"id":"4e7027fc-d7e6-4cf6-876d-3589674dd551","name":"lighthouse","color":"gray"},{"id":"25628f01-b484-47bc-96b1-9c2e0b028e2a","name":"core web vitals","color":"purple"},{"id":"54d6d94e-19ca-405d-8585-2902f3cb6d1f","name":"portfolio","color":"default"}],"publishDate":"2022-03-14T05:24:00.000Z"}]
