+++ src/util/social.ts
@@ -1,0 +1,53 @@
+import { FaCameraRetro } from "react-icons/fa";
+import { SiDevdotto, SiGithub, SiGmail, SiLinkedin } from "react-icons/si";
+
+export const quicklinks = [
+    {
+        Icon: SiGithub,
+        text: "Github",
+        href: "https://github.com/TheForeverLost",
+    },
+    {
+        Icon: SiLinkedin,
+        text: "Linkedin",
+        href: "https://www.linkedin.com/in/alan-john-b2b521193/",
+    },
+    {
+        Icon: FaCameraRetro,
+        text: "Pictures",
+        href: "https://photos.alanjohn.dev",
+    },
+    {
+        Icon: SiDevdotto,
+        text: "DevTo",
+        href: "https://dev.to/theforeverlost",
+    },
+    {
+        Icon: SiGmail,
+        text: "Gmail",
+        href: "mailto:alansandra2013@gmail.com",
+    },
+];
+
+export const navigationLinkData = [
+    {
+        text: "Home",
+        href: "/",
+    },
+    {
+        text: "About",
+        href: "/#About",
+    },
+    {
+        text: "Projects",
+        href: "/#Projects",
+    },
+    {
+        text: "Blog",
+        href: "/blog",
+    },
+    {
+        text: "Photos",
+        href: "https://photos.alanjohn.dev",
+    },
+];
